backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
boto3==1.35.40
python-jose[cryptography]==3.3.0
httpx==0.27.2
redis==5.0.8
rq==1.16.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
psycopg[binary]==3.2.3
orjson==3.10.7
langdetect==1.0.9
numpy==2.1.2
onnxruntime==1.19.2
fakeredis==2.23.2

